requirements.txt
# Development
django-debug-toolbar>=1.2.1,<1.3
django-debug-toolbar-template-timings==0.6.1
django-extensions>=1.3.3,<1.4.0
Werkzeug>=0.9.4,<0.10

# Sandbox
Whoosh>=2.5
pysolr>=3.1
requests==2.1
uWSGI==1.9.21.1

# Docs
Sphinx==1.2.2
sphinxcontrib-napoleon==0.2.7
sphinx_rtd_theme==0.1.6

# Testing
django-dynamic-fixture>=1.7,<1.8
mock>=1.0.1,<1.1
coverage>=3.6,<3.7
django-nose>=1.2,<1.3
spec>=0.11.1,<0.12
WebTest>=2.0,<2.1
django-webtest>=1.7,<1.8
tox==1.6.1
coveralls>=0.1.1,<0.2
purl>=0.4
behave==1.2.3
factory-boy==2.3.1

# Misc
flake8>=2.1
pyprof2calltree==1.3.1
ipdb>=0.8,<0.9
ipython>=1.1.0,<1.2.0

# House of Haus Additions
django-allauth>=0.17.0
django-endless-pagination
django-relationships==0.3.0
django-oscar-accounts

